i+ and i- cases for the normal case, still some issues for the navier stokes case and a *lot* of refactoring

+++ i0/navier_stokes_extra_test.go
@@ -9,7 +9,7 @@ func Test_ns_pn(t *testing.T) {
 	tests := []struct {
 		name  string
 		p, q  Vector
-		debth int
+		debth int64
 		want  Vector
 	}{
 		{
